1단계. Splitter 를 OperandFilter 로 변경
